Start moving shared code into one project
