--- UART-CL By TheCod3r/.idea/.idea.UART-CL By TheCod3r/.idea/.gitignore
@@ -1,13 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
-# Rider ignored files
-/contentModel.xml
-/modules.xml
-/projectSettingsUpdater.xml
-/.idea.UART-CL By TheCod3r.iml
-# Editor-based HTTP Client requests
-/httpRequests/
-# Datasource local storage ignored files
-/dataSources/
-/dataSources.local.xml
