--- UART-CL By TheCod3r/UART-CL By TheCod3r.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.3.32901.215
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UART-CL By TheCod3r", "UART-CL By TheCod3r\UART-CL By TheCod3r.csproj", "{8B5DA648-23A0-4C54-A0D2-F6B5E326F308}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8B5DA648-23A0-4C54-A0D2-F6B5E326F308}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8B5DA648-23A0-4C54-A0D2-F6B5E326F308}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8B5DA648-23A0-4C54-A0D2-F6B5E326F308}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8B5DA648-23A0-4C54-A0D2-F6B5E326F308}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {BAD6A308-3F90-4873-A39E-365E07239E9B}
-	EndGlobalSection
-EndGlobal
