--- UART-CL By TheCod3r/UART-CL By TheCod3r/Properties/PublishProfiles/ClickOnceProfile.pubxml
@@ -1,39 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<!--
-https://go.microsoft.com/fwlink/?LinkID=208121.
--->
-<Project>
-  <PropertyGroup>
-    <ApplicationRevision>2</ApplicationRevision>
-    <ApplicationVersion>1.1.0.*</ApplicationVersion>
-    <BootstrapperEnabled>True</BootstrapperEnabled>
-    <Configuration>Release</Configuration>
-    <CreateDesktopShortcut>True</CreateDesktopShortcut>
-    <CreateWebPageOnPublish>False</CreateWebPageOnPublish>
-    <GenerateManifests>true</GenerateManifests>
-    <Install>True</Install>
-    <InstallFrom>Disk</InstallFrom>
-    <IsRevisionIncremented>True</IsRevisionIncremented>
-    <IsWebBootstrapper>False</IsWebBootstrapper>
-    <MapFileExtensions>True</MapFileExtensions>
-    <OpenBrowserOnPublish>False</OpenBrowserOnPublish>
-    <Platform>Any CPU</Platform>
-    <ProductName>UART-CL</ProductName>
-    <PublishDir>bin\Release\net6.0\publish\</PublishDir>
-    <PublishUrl>bin\Release\net6.0\publish\</PublishUrl>
-    <PublisherName>TheCod3r</PublisherName>
-    <PublishProtocol>ClickOnce</PublishProtocol>
-    <PublishReadyToRun>False</PublishReadyToRun>
-    <PublishSingleFile>True</PublishSingleFile>
-    <RuntimeIdentifier>win-x64</RuntimeIdentifier>
-    <SelfContained>True</SelfContained>
-    <SignatureAlgorithm>(none)</SignatureAlgorithm>
-    <SignManifests>False</SignManifests>
-    <TargetFramework>net6.0</TargetFramework>
-    <UpdateEnabled>False</UpdateEnabled>
-    <UpdateMode>Foreground</UpdateMode>
-    <UpdateRequired>False</UpdateRequired>
-    <WebPageFileName>Publish.html</WebPageFileName>
-    <History>True|2024-04-28T21:23:14.6592026Z;</History>
-  </PropertyGroup>
-</Project>
