--- UART-CL By TheCod3r/UART-CL By TheCod3r/Properties/PublishProfiles/ClickOnceProfile.pubxml.user
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<!--
-https://go.microsoft.com/fwlink/?LinkID=208121.
--->
-<Project>
-  <PropertyGroup>
-    <History>True|2024-04-28T21:28:11.6019250Z;True|2024-04-28T22:23:14.6592026+01:00;</History>
-    <LastFailureDetails />
-  </PropertyGroup>
-</Project>
