--- UART-CL By TheCod3r/UART-CL By TheCod3r/Properties/PublishProfiles/FolderProfile1.pubxml.user
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<!--
-https://go.microsoft.com/fwlink/?LinkID=208121.
--->
-<Project>
-  <PropertyGroup>
-    <History>True|2024-04-28T21:03:32.2559262Z;True|2024-04-27T00:21:15.8137671+01:00;</History>
-    <LastFailureDetails />
-  </PropertyGroup>
-</Project>
