--- UART-CL By TheCod3r/UART-CL By TheCod3r/Properties/PublishProfiles/FolderProfile2.pubxml.user
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<!--
-https://go.microsoft.com/fwlink/?LinkID=208121.
--->
-<Project>
-</Project>
