--- UART-CL By TheCod3r/UART-CL By TheCod3r/UART-CL By TheCod3r.csproj.user
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <_LastSelectedProfileId>C:\Users\TheCod3r\source\repos\UART-CL By TheCod3r\UART-CL By TheCod3r\Properties\PublishProfiles\ClickOnceProfile.pubxml</_LastSelectedProfileId>
-  </PropertyGroup>
-</Project>
